+++ src/agh/to/carfactory/Option.java
@@ -3,7 +3,7 @@ package agh.to.carfactory;
 public class Option extends CarComponent{
     private String optionName;
 
-    public Option(String optionName, int additionalFee) {
+    public Option(String optionName, double additionalFee) {
         super(additionalFee);
         this.optionName = optionName;
     }
